backend/requirements.txt
fastapi
uvicorn
python-multipart
llama-index
chromadb
python-dotenv
llama-index-llms-gemini
google-generativeai
nest-asyncio
